phy: share the chip-independent test scaffolding across fixtures
Delayer, the no-op DummyVariant and the never-failing SPI error type
were byte-identical copies in all three comparison fixtures (and the
emulator carried a fourth error type). They now live in one
cfg(test) module. Each fixture's transaction recording and equality
rules stay put — those are the chip-specific, load-bearing parts.
