+++ lora-phy/src/lib.rs
@@ -33,6 +33,8 @@ pub mod mod_traits;
 pub mod sx126x;
 /// Specific implementation to support Semtech Sx127x chips
 pub mod sx127x;
+#[cfg(test)]
+pub(crate) mod test_fixtures;
 
 pub use crate::mod_params::RxMode;
