+++ lora-phy/src/sx126x/test/emulator.rs
@@ -5,7 +5,8 @@ use crate::mod_params::RadioError;
 use crate::mod_traits::InterfaceVariant;
 use crate::sx126x::radio_kind_params::{IrqMask, OpCode};
 use crate::sx126x::{Config, Sx1261, Sx126x};
-use embedded_hal::spi::{ErrorKind, Operation};
+use crate::test_fixtures::SpiError;
+use embedded_hal::spi::Operation;
 use embedded_hal_async::delay::DelayNs;
 use embedded_hal_async::spi::SpiDevice;
 use std::collections::HashMap;
@@ -242,14 +243,6 @@ impl Chip {
 
     pub fn with_model<R>(&self, f: impl FnOnce(&mut ChipModel) -> R) -> R {
         f(&mut self.0.lock().unwrap())
-    }
-}
-
-#[derive(Debug)]
-pub enum SpiError {}
-impl embedded_hal::spi::Error for SpiError {
-    fn kind(&self) -> ErrorKind {
-        todo!()
     }
 }
