+++ lora-phy/src/sx127x/test/fixtures.rs
@@ -1,8 +1,6 @@
-use crate::mod_params::RadioError;
-use crate::mod_traits::InterfaceVariant;
 use crate::sx127x::{Config, Sx1276, Sx127x};
-use embedded_hal::spi::{ErrorKind, Operation};
-use embedded_hal_async::delay::DelayNs;
+pub use crate::test_fixtures::{Delayer, DummyVariant, SpiError};
+use embedded_hal::spi::Operation;
 use embedded_hal_async::spi::SpiDevice;
 use std::collections::{BTreeMap, VecDeque};
 
@@ -171,15 +169,8 @@ impl TestFixture {
     }
 }
 
-#[derive(Debug)]
-pub enum Error {}
-impl embedded_hal::spi::Error for Error {
-    fn kind(&self) -> ErrorKind {
-        todo!()
-    }
-}
 impl embedded_hal::spi::ErrorType for TestFixture {
-    type Error = Error;
+    type Error = SpiError;
 }
 
 impl embedded_hal::spi::SpiDevice for TestFixture {
@@ -240,32 +231,4 @@ pub fn get_sx1276_boost() -> Sx127x<TestFixture, DummyVariant, Sx1276> {
             rx_boost: false,
         },
     )
-}
-
-pub struct Delayer;
-impl DelayNs for Delayer {
-    async fn delay_ns(&mut self, _ns: u32) {}
-}
-
-pub struct DummyVariant;
-
-impl InterfaceVariant for DummyVariant {
-    async fn reset(&mut self, _delay: &mut impl DelayNs) -> Result<(), RadioError> {
-        Ok(())
-    }
-    async fn wait_on_busy(&mut self) -> Result<(), RadioError> {
-        Ok(())
-    }
-    async fn await_irq(&mut self) -> Result<(), RadioError> {
-        Ok(())
-    }
-    async fn enable_rf_switch_rx(&mut self) -> Result<(), RadioError> {
-        Ok(())
-    }
-    async fn enable_rf_switch_tx(&mut self) -> Result<(), RadioError> {
-        Ok(())
-    }
-    async fn disable_rf_switch(&mut self) -> Result<(), RadioError> {
-        Ok(())
-    }
 }
